refactor: extract messageCreate handler
